TMBF: Scale DAC gain control to friendly range +-1
git-svn-id: svn+ssh://serv2/home/subversion/repos/controls/diamond/trunk/ioc/TMBF@115009 e099a375-04f9-0310-9d5f-a741eaff62e1

+++ tmbfApp/Db/bunch_select.py
@@ -18,7 +18,7 @@ for bank in range(4):
     # Waveform settings with status update
     BunchWaveforms(bank, 'FIRWF', 'CHAR', 'FIR bank select')
     BunchWaveforms(bank, 'OUTWF', 'CHAR', 'DAC output select')
-    BunchWaveforms(bank, 'GAINWF', 'LONG', 'DAC output gain')
+    BunchWaveforms(bank, 'GAINWF', 'FLOAT', 'DAC output gain')
 
 
 # Bunch synchronisation.
